+++ src/ScopelyBattles.Shared/Battles/StolenResources.cs
@@ -4,8 +4,8 @@ public sealed record StolenResources
 {
     public required int GoldPercent { get; init; }
     public required int SilverPercent { get; init; }
-    public required int Gold { get; init; }
-    public required int Silver { get; init; }
+    public required long Gold { get; init; }
+    public required long Silver { get; init; }
 
-    public int Total => Gold + Silver;
+    public long Total => Gold + Silver;
 }
